made health int

+++ src/game/game.c
@@ -163,7 +163,7 @@ f32 game_get_player_health_ratio(void)
     if (player.entity == NULL || game_paused)
         return 0;
     assert(player.entity != NULL);
-    return player.entity->health / player.entity->max_health;
+    return (f32)player.entity->health / player.entity->max_health;
 }
 
 u32 game_get_weapon_tex(void)
